Fixes from comments
